+++ src/main/java/tonius/simplyjetpacks/config/Defaults.java
@@ -8,7 +8,6 @@ public abstract class Defaults {
     // Item
     public static final boolean enableFuelEfficiencyEnchantment = true;
     public static final boolean addRAItemsIfNotInstalled = true;
-
     // Integration
     public static final boolean enableIntegrationVanilla = true;
     public static final boolean enableIntegrationEIO = ModType.ENDER_IO.loaded;
@@ -19,20 +18,16 @@ public abstract class Defaults {
     public static boolean enableIntegrationRA = ModType.REDSTONE_ARSENAL.loaded;
     public static final boolean enableIntegrationRR = ModType.REDSTONE_REPOSITORY.loaded;
     public static final int gelidEnderiumFuelUsageBonus = 80;
-
     // Controls
     public static final boolean customControls = false;
     public static final String flyKey = "SPACE";
     public static final String descendKey = "LSHIFT";
     public static final boolean invertHoverSneakingBehavior = false;
     public static final boolean doubleTapSprintInAir = true;
-
     // Aesthetics
     public static final boolean enableArmor3DModels = true;
-
     // Sounds
     public static final boolean jetpackSounds = true;
-
     // GUI
     public static final boolean holdShiftForDetails = true;
     public static final HUDPositions HUDPosition = HUDPositions.TOP_LEFT;