+++ src/main/java/tonius/simplyjetpacks/gui/element/ElementEnergyStoredAdv.java
@@ -5,6 +5,6 @@ public class ElementEnergyStoredAdv { //extends ElementEnergyStored {
     // TODO: Re-add GUIs
     /*public ElementEnergyStoredAdv(GuiBase gui, int posX, int posY, IEnergyStorage storage) {
         super(gui, posX, posY, storage);
-        this.texture = new ResourceLocation(SimplyJetpacks.RESOURCE_PREFIX + "textures/gui/elements/bar_energy.png");
+        this.texture = new ResourceLocation(SimplyJetpacks.RESOURCE_PREFIX + "textures/gui/elements/energy_bar.png");
     }*/
 }